+++ .expo/packager-info.json
@@ -1,10 +1,10 @@
 {
   "devToolsPort": 19002,
   "expoServerPort": 19000,
-  "packagerPort": 19001,
-  "packagerPid": 4291,
+  "packagerPort": 19003,
+  "packagerPid": 14688,
   "expoServerNgrokUrl": "https://sz-iug.anonymous.mvp-map.exp.direct",
   "packagerNgrokUrl": "https://packager.sz-iug.anonymous.mvp-map.exp.direct",
-  "ngrokPid": 4316,
-  "webpackServerPort": null
+  "ngrokPid": 14711,
+  "webpackServerPort": 19006
 }
