+++ package.json
@@ -18,7 +18,10 @@
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
     "react-native-vector-icons": "^6.6.0",
-    "react-native-web": "~0.11.7"
+    "react-native-web": "~0.11.7",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5",
+    "redux-axios-middleware": "^4.0.1"
   },
   "devDependencies": {
     "babel-preset-expo": "~8.0.0",
