+++ store/reducers/locationReducers.js
@@ -1,0 +1,17 @@
+const initialState = null;
+
+const locationReducers = (state = initialState, { type, payload }) => {
+  switch (type) {
+    case 'SET_LOCATION':
+      return (
+        {
+          ...state,
+          location: payload,
+        }
+      );
+    default:
+      return state;
+  }
+};
+
+export default locationReducers;
